spectator list !!!

+++ cs2-internal/sdk/entity/CPlayerObserverServices.hpp
@@ -3,6 +3,6 @@
 
 class CPlayerObserverServices {
 public:
-	SCHEMA_FIELD_OFFSET(i32, m_iObserverMode, 0x40);
+	SCHEMA_FIELD_OFFSET(u8, m_iObserverMode, 0x40);
 	SCHEMA_FIELD_OFFSET(CHandle, m_hObserverTarget, 0x44);
 };
